refactor AiChat to use AiChatBase and rename it to AiChatMarket
